requirements.txt
Django==4.2.7
google-generativeai==0.3.2
PyMuPDF==1.23.8
fpdf2==2.7.6
python-dotenv==1.0.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-allauth==0.57.0
django-environ==0.11.2
whitenoise==6.6.0
gunicorn==21.2.0

